refactor: add persistence repositories and developer docs (#61)

+++ binance_trade_bot/models/deposit.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, String, Text
+from sqlalchemy import Column, DateTime, Float, Integer, String, Text
 
 from .base import Base
 
@@ -9,7 +9,7 @@ class Deposit(Base):
     """Record of capital deposited into the trading account."""
     __tablename__ = "deposits"
 
-    id = Column(Float, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     amount = Column(Float, nullable=False)
     currency = Column(String, default="USDC")
     source = Column(String, default="manual")
